Tidy target-commits internals
Move the target-commits endpoint and its transport types out of the
oversized workspace module into their own module, and key the review
lookup by object id instead of re-encoding hex per commit.

In the upstream outline, derive the workspace boundary without mutating
a captured variable — that pattern made React Compiler bail on the whole
hook, so the combined listing was rebuilt with fresh identities on every
render. Share one older-history query hook between the outline and the
show-more row, skip refetching loaded history pages on local workspace
activity since only a fetch can move them, and simplify how the item
builder accumulates branch stubs.
